edit project with review comments in pull request
