+++ Phase01-Search/src/main/java/FileSaving.java
@@ -2,11 +2,11 @@ import java.io.File;
 
 public interface FileSaving {
 
-     static void addAllFiles(InvertedIndex invertedIndex) {
+     static void addAllFiles(HashInvertedIndex invertedIndex) {
 
     }
 
-     static void separate(File file, InvertedIndex invertedIndex) {
+     static void separate(File file, HashInvertedIndex invertedIndex) {
 
     }
 }
