--- Phase01-Search/src/main/java/InvertedIndexAbility.java
@@ -1,5 +1,0 @@
-import java.util.ArrayList;
-
-public interface InvertedIndexAbility {
-    public ArrayList<String> findDoc(String word);
-}
