+++ Phase01-Search/src/main/java/StringSeparation.java
@@ -4,9 +4,5 @@ public interface StringSeparation {
 
     public String normalize(String word);
 
-    public ArrayList<String> findNothingWords(String[] splitInput);
-
-    public ArrayList<String> findPlusWords(String[] splitInput);
-
-    public ArrayList<String> findMinusWords(String[] splitInput);
+    public ArrayList<String> findWordsFromInput(String[] splitInput, String regex);
 }
